Update bmi08a.c
Fix accelerometer self-test.  Adding the absolute values of the readings together does not compute the difference.

I found that accel_pos readings are positive and accel_neg readings are negative, so it could be simplified to `accel_pos->x + accel_neg->x`, but taking the absolute value of the difference will work either way.